Title: 그 전 알파벳, Time: 118ms, Memory: 10MB, Status: Passed - CodeTree

+++ 241209/그 전 알파벳/before-the-alphabet.java
@@ -10,7 +10,7 @@ public class Main {
         if (ch == 'a')
             result = 'z';
         else
-            result = (char)((int)ch - 1);
+            result = (char)(ch - 1);
 
         System.out.println(result);
     }
